requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery>=5.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==2.3.0
Django<5.0,>=4.2
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.0.0
django-markdown-deux==1.0.6
django-redis==5.4.0
django-tailwind==4.0.1
python-dotenv>=1.0.0
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn>=21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-text-splitters==0.3.8
langsmith==0.3.41
Markdown>=3.4.0
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy>=1.24.0
openai>=1.0.0
orjson==3.10.18
packaging==24.2
Pillow>=10.0.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary>=2.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis>=4.5.0
requests>=2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
zstandard==0.23.0
markdown==3.5.1
bleach==6.1.0
pygments==2.17.2
twilio>=8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
